Settings: Separate Expanded desktop from Power menu settings
This commit separates the ability to configure the Expanded desktop
mode from the ability to enable the switch in the Power menu.  This
will allow for the ED Tile to be hidden if the feature is disabled
and the power menu entry to be removed if the user chooses to use the
Tile instead.

Change-Id: I3d2eabde8087d25940cbb1e5262086b9c6e291f1

Conflicts:
	res/xml/system_settings.xml
	src/com/android/settings/cyanogenmod/SystemSettings.java